requirements.txt
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.0
requests==2.31.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
coreapi==2.3.3
python-dotenv==1.0.0 